Changed the way the icon font is added.
Forced the Icon font to first check if the current system has the same font installed. If so, use the system font. And use the embedded font otherwise. This makes the designer load the icon font properly at design time.

This will be reverted prior to publish the source code!