Splitted the solution into two projects.
Project 1 (Morgan.Core)
-------------------------
A .NET Standard project that contains the base of the complete PROJECT (as a whole, including both mobile and desktop apps), such as DataModels, services, VMs etc. This project is aimed to be cross-platform in order to be usable with Xamarin-Forms.

Project 2 (Morgan)
--------------------
Contains everything related to the WPF Windows application. Required services are registered in the App.xaml.cs file by overridding the Startup method.